expose more code in examples
