+++ docs/src/lib/registry/examples/message-scroller-opening-position.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
+	import OpeningPositionScroller from "$lib/components/message-scroller/opening-position-scroller.svelte";
 	import * as Card from "$lib/registry/ui/card/index.js";
 	import * as MessageScroller from "$lib/registry/ui/message-scroller/index.js";
 	import * as Tabs from "$lib/registry/ui/tabs/index.js";
-	import OpeningPositionScroller from "$lib/components/message-scroller/opening-position-scroller.svelte";
 
 	const positions = [
 		{ value: "start", label: "start" },
@@ -13,8 +13,8 @@
 		label: string;
 	}>;
 
-	let positionKey = $state(0);
 	let position = $state<"start" | "end" | "last-anchor">("last-anchor");
+	let positionKey = $state(0);
 
 	function handlePositionChange(value: string) {
 		if (value === "start" || value === "end" || value === "last-anchor") {
